+++ App.tsx
@@ -1,37 +1,19 @@
 import { NavigationContainer } from '@react-navigation/native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
 import { ChatScreen } from './screens/ChatScreen/ChatScreen';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { Provider } from 'react-redux';
 import { store } from './store/store';
+import { RootNavigator } from './components/RootNavigator/RootNavigator';
 
-const Stack = createNativeStackNavigator();
 
 export default function App() {
   return (
     <Provider store={store}>
       <GestureHandlerRootView style={{flex: 1}}>
         <NavigationContainer>
-          <Stack.Navigator>
-            <Stack.Screen 
-              name='ChatScreen'
-              component={ChatScreen}
-              options={{headerShown: false}}
-            />
-          </Stack.Navigator>
+          <RootNavigator/>
         </NavigationContainer>
       </GestureHandlerRootView>
     </Provider>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
